Fix async usage
