+++ package.json
@@ -2,7 +2,7 @@
   "name": "@resistbot/middy-beeline",
   "version": "1.0.0",
   "description": "Middy Middleware for Honeycomb Beeline",
-  "main": "index.js",
+  "main": "src/index.js",
   "engines": {
     "node": ">=12.13.0"
   },
@@ -32,6 +32,7 @@
     "honeycomb-beeline": "^2.7.4"
   },
   "devDependencies": {
+    "@serverless/event-mocks": "^1.1.1",
     "eslint": "^7.32.0",
     "eslint-config-airbnb-base": "^14.2.1",
     "eslint-config-prettier": "^8.3.0",
