+++ src/__mocks__/honeycomb-beeline.js
@@ -1,0 +1,7 @@
+const configure = () => {
+  return configure;
+};
+
+configure.startTrace = () => 'FAKE TRACE';
+
+module.exports = configure;
